fix: renamed indicator and reduced its opacity to match os
